move: AuthorizedService 구현체 이동
