+++ src/main/java/codesquad/fineants/global/common/authorized/service/AuthorizedService.java
@@ -1,4 +1,4 @@
-package codesquad.fineants.global.common.authorized;
+package codesquad.fineants.global.common.authorized.service;
 
 import java.util.List;
